Add `CyclotomicField` to represent cyclotomic operations

+++ ff/src/biginteger/arithmetic.rs
@@ -83,7 +83,7 @@ pub(crate) fn mac_with_carry(a: u64, b: u64, c: u64, carry: &mut u64) -> u64 {
 }
 
 /// Compute the NAF (non-adjacent form) of num
-pub fn find_naf(num: &[u64]) -> Vec<i64> {
+pub fn find_naf(num: &[u64]) -> Vec<i8> {
     let is_zero = |num: &[u64]| num.iter().all(|x| *x == 0u64);
     let is_odd = |num: &[u64]| num[0] & 1 == 1;
     let sub_noborrow = |num: &mut [u64], z: u64| {
@@ -118,9 +118,9 @@ pub fn find_naf(num: &[u64]) -> Vec<i64> {
     let mut res = vec![];
 
     while !is_zero(&num) {
-        let z: i64;
+        let z: i8;
         if is_odd(&num) {
-            z = 2 - (num[0] % 4) as i64;
+            z = 2 - (num[0] % 4) as i8;
             if z >= 0 {
                 sub_noborrow(&mut num, z as u64)
             } else {
@@ -147,9 +147,9 @@ pub fn find_naf(num: &[u64]) -> Vec<i64> {
 /// This can be rewritten as ...0 1 1 to avoid one doubling, at the cost that we are making an
 /// exception of non-adjacence for the most significant bit.
 ///
-/// Since this representation is no longer a strict NAF, we call it ``relaxed NAF''.
+/// Since this representation is no longer a strict NAF, we call it "relaxed NAF".
 ///
-pub fn find_relaxed_naf(num: &[u64]) -> Vec<i64> {
+pub fn find_relaxed_naf(num: &[u64]) -> Vec<i8> {
     let mut res = find_naf(num);
 
     let len = res.len();
